python/requirements.txt
requests==2.30.0
playwright
sqlalchemy
psycopg2-binary
python-dotenv
